// src/app/api/model/savingsGoal.ts
/**
 * Personal Finance Management API
 *
 * Contact: dragisa@yahoo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SavingsGoal { 
    id?: number;
    name?: string;
    targetAmount?: number;
    currentAmount?: number;
    targetDate?: string;
    completed?: boolean;
    username?: string;
    createdAt?: string;
    updatedAt?: string;
}